Create bootable sector
This is the simplest safe program that can realy be created.

This creates a 512 bit sector ending in 0xaa55; intel's magic number.

The loop is created to stop the computer from continuing reading
sectors. It could be dangerous to keep reading because there could
theoretically be valid commands stored in any random binary file.