Implement hello world
As long as we stay in 16 bit real mode the ah register is used by bios
interrupts to choose a mode, and the al register contains the arguments
for teh interupt. when ah=0x0e the mode is scrolling teletype, and al
contains the character to print to screen when the interrupt 0x10
happens.